chore: move `build_extended` logic to lib
