+++ cargo-near-build/src/lib.rs
@@ -15,6 +15,9 @@ pub use types::near::build::Opts as BuildOpts;
 
 pub use types::near::build::version_mismatch::VersionMismatch;
 pub use types::near::build::CompilationArtifact as BuildArtifact;
+pub use types::near::build_extended::{
+    build_script::Opts as BuildScriptOpts, OptsExtended as BuildOptsExtended,
+};
 
 pub use near_abi;
 // used in `AbiOpts` and `BuildOpts`, and `BuildArtifact`
@@ -22,3 +25,4 @@ pub use camino;
 
 pub use near::abi::build as build_abi;
 pub use near::build::run as build;
+pub use near::build_extended::run as build_extended;
