+++ cargo-near-build/src/near/build_extended/mod.rs
@@ -1,19 +1,10 @@
 mod build_script;
 mod tmp_env;
-pub use build_script::BuildScriptOpts;
-use cargo_near_build::{BuildArtifact, BuildOpts};
+use crate::types::near::build::CompilationArtifact;
+use crate::types::near::build_extended::OptsExtended;
 use rustc_version::Version;
 
-#[derive(Debug, Clone)]
-pub struct OptsExtended<'a> {
-    pub workdir: &'a str,
-    /// vector of key-value pairs of temporary env overrides during build process
-    pub env: Vec<(&'a str, &'a str)>,
-    pub build_opts: BuildOpts,
-    pub build_script_opts: BuildScriptOpts<'a>,
-}
-
-pub fn build(args: OptsExtended) -> Result<BuildArtifact, Box<dyn std::error::Error>> {
+pub fn run(args: OptsExtended) -> Result<CompilationArtifact, Box<dyn std::error::Error>> {
     let actual_version = rustc_version::version()?;
     let (artifact, skipped) = args.skip_or_compile(&actual_version)?;
 
@@ -26,7 +17,7 @@ impl<'a> OptsExtended<'a> {
     pub fn skip_or_compile(
         &self,
         version: &Version,
-    ) -> Result<(BuildArtifact, bool), Box<dyn std::error::Error>> {
+    ) -> Result<(CompilationArtifact, bool), Box<dyn std::error::Error>> {
         let _tmp_workdir = tmp_env::set_current_dir(self.workdir)?;
         let result = if self.build_script_opts.should_skip(version) {
             let artifact = self.build_script_opts.create_empty_stub()?;
@@ -40,7 +31,7 @@ impl<'a> OptsExtended<'a> {
 
     /// `CARGO_TARGET_DIR` export is needed to avoid attempt to acquire same `target/<profile-path>/.cargo-lock`
     /// as the `cargo` process, which is running the build-script
-    pub fn compile_near_artifact(&self) -> Result<BuildArtifact, Box<dyn std::error::Error>> {
+    pub fn compile_near_artifact(&self) -> Result<CompilationArtifact, Box<dyn std::error::Error>> {
         let mut tmp_envs = vec![];
         for (env_key, value) in self.env.iter() {
             let tmp_env = tmp_env::set_var(env_key, value);
@@ -51,9 +42,9 @@ impl<'a> OptsExtended<'a> {
         {
             let _tmp_cargo_target_env = tmp_env::set_var("CARGO_TARGET_DIR", distinct_target_dir);
 
-            cargo_near_build::build(self.build_opts.clone())?
+            crate::build(self.build_opts.clone())?
         } else {
-            cargo_near_build::build(self.build_opts.clone())?
+            crate::build(self.build_opts.clone())?
         };
 
         Ok(artifact)
