+++ cargo-near-build/src/near/mod.rs
@@ -1,3 +1,4 @@
 // TODO: make mod non-pub
 pub mod abi;
 pub mod build;
+pub mod build_extended;
