--- cargo-near-build/src/types/near.rs
@@ -1,3 +1,0 @@
-// TODO: make non-pub
-pub mod abi;
-pub mod build;
