+++ cargo-near/Cargo.toml
@@ -28,7 +28,6 @@ clap = { version = "4.0.18", features = ["derive", "env"] }
 colored = "2.0"
 env_logger = "0.11.5"
 log = "0.4"
-rustc_version = "0.4"
 serde = "1.0.197"
 serde_json = "1.0"
 color-eyre = "0.6"
